Revert "replace source region floats with doubles to handle small fluxes appropriately"
This reverts commit aea1f179b1beb462df5bf07bce9478af07521d70.

+++ include/openmc/random_ray/random_ray.h
@@ -70,17 +70,17 @@ public:
 
   //----------------------------------------------------------------------------
   // Public data members
-  vector<double> angular_flux_;
+  vector<float> angular_flux_;
   bool ray_trace_only_ {false}; // If true, only perform geometry operations
 
   //---------------------------------------------------------------------------
   // Public data members for kinetic simulations
-  vector<double> angular_flux_prime_;
+  vector<float> angular_flux_prime_;
 
 private:
   //----------------------------------------------------------------------------
   // Private data members
-  vector<double> delta_psi_;
+  vector<float> delta_psi_;
   vector<MomentArray> delta_moments_;
   vector<int> mesh_bins_;
   vector<double> mesh_fractional_lengths_;
